+++ #/handlers/pos/product/uom-editor.php
@@ -1,5 +1,13 @@
 <?php
 
+function _ensure_user_can_access() {
+  global $uom;
+  if ($uom->id == 0 && !current_user_can('add-product-uom'))
+    exit(render('error/403'));
+  else if ($uom->id != 0 && !current_user_can('edit-product-uom'))
+    exit(render('error/403'));
+}
+
 require_once CORELIB_PATH . '/Product.php';
 
 $id = (int)(isset($_REQUEST['id']) ? $_REQUEST['id'] : 0);
@@ -25,15 +33,20 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
   $action = $_POST['action'];
   
   if ($action == 'delete') {
+    if (!current_user_can('delete-product-uom'))
+      exit(render('error/403'));
+      
     $db->query('delete from product_uoms where id=' . $uom->id);
     $_SESSION['FLASH_MESSAGE'] = "Satuan $uom->name telah dihapus.";
     header('Location: ./editor?id=' . $uom->productId);
     exit;
   }
   
+  _ensure_user_can_access();
+  
   if ($action == 'save') {
-    $uom->name = (string)$_POST['name'];
-    $uom->quantity = (int)str_replace('.', '', (string)$_POST['quantity']);
+    $uom->name = trim((string)$_POST['name']);
+    $uom->quantity = (int)str_replace('.', '', trim((string)$_POST['quantity']));
     
     if (!$uom->id) {
       $q = $db->prepare('insert into product_uoms( productId, name, quantity) values (:productId,:name,:quantity)');
@@ -52,6 +65,8 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
     exit;
   }
 }
+
+_ensure_user_can_access();
 
 render('pos/product/uom-editor', [
   'uom'     => $uom,
